# test-api-server/openapi.yaml
openapi: 3.0.3
info:
  title: MCP Test API Server
  description: |
    A comprehensive API server for testing MCP Studio functionality. 
    Provides CRUD operations for users, chats, messages, and feedbacks with MongoDB backend.

    This API is designed to be easily convertible to an MCP (Model Context Protocol) server.
  version: 1.0.0
  contact:
    name: MCP Studio Team
    email: support@mcpstudio.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.mcpstudio.com
    description: Production server
...
paths:
...
  /api/users:
    get:
      tags: [users]
      summary: Get all users
      description: Retrieve a paginated list of all users
      operationId: getUsers
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
        "500":
          description: Failed to fetch users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags: [users]
      summary: Create new user
      description: Create a new user with email address
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateUserResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to create user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/users/{id}:
    get:
      tags: [users]
      summary: Get user by ID
      description: Retrieve a specific user by their ID
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          description: User ID (ObjectId)
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      responses:
        "200":
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid user ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to fetch user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      tags: [users]
      summary: Update user
      description: Update an existing user's information
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          description: User ID (ObjectId)
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateResponse"
        "400":
          description: Invalid input or user ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to update user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [users]
      summary: Delete user
      description: Delete a user by their ID
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          description: User ID (ObjectId)
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      responses:
        "200":
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "400":
          description: Invalid user ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to delete user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/users/email/{email}:
    get:
      tags: [users]
      summary: Get user by email
      description: Retrieve a specific user by their email address
      operationId: getUserByEmail
      parameters:
        - name: email
          in: path
          required: true
          description: User email address
          schema:
            type: string
            format: email
      responses:
        "200":
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to fetch user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # Chat endpoints
  /api/chats:
    get:
      tags: [chats]
      summary: Get all chats
      description: Retrieve a paginated list of all chats
      operationId: getChats
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Chats retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatListResponse"
        "500":
          description: Failed to fetch chats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags: [chats]
      summary: Create new chat
      description: Create a new chat for a user
      operationId: createChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChatRequest"
      responses:
        "201":
          description: Chat created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateChatResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Chat already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to create chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/chats/user/{userId}:
    get:
      tags: [chats]
      summary: Get chats by user ID
      description: Retrieve all chats owned by a specific user
      operationId: getChatsByUserId
      parameters:
        - name: userId
          in: path
          required: true
          description: User ID (ObjectId)
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: User chats retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatListResponse"
        "400":
          description: Invalid user ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to fetch user chats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/messages:
    get:
      tags: [messages]
      summary: Get all messages
      description: Retrieve a paginated list of all messages
      operationId: getMessages
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Messages retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageListResponse"
        "500":
          description: Failed to fetch messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags: [messages]
      summary: Create message
      description: Create a new message in a chat
      operationId: createMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageRequest"
      responses:
        "201":
          description: Message created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateMessageResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to create message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/feedbacks:
    get:
      tags: [feedbacks]
      summary: Get all feedbacks
      description: Retrieve a paginated list of all feedbacks
      operationId: getFeedbacks
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Feedbacks retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackListResponse"
        "500":
          description: Failed to fetch feedbacks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/feedbacks/user/{userId}:
    get:
      tags: [feedbacks]
      summary: Get user feedback history
      description: Retrieve feedback history for a specific user with filtering and sorting
      operationId: getUserFeedbackHistory
      parameters:
        - name: userId
          in: path
          required: true
          description: User ID (ObjectId)
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: rating
          in: query
          description: Filter by rating
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: isUpvoted
          in: query
          description: Filter by upvote status
          schema:
            type: boolean
        - name: hasJiraIssue
          in: query
          description: Filter by Jira issue presence
          schema:
            type: boolean
        - name: sortBy
          in: query
          description: Sort field
          schema:
            type: string
            enum: [createdAt, rating, lastModifiedAt]
            default: createdAt
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: User feedback history retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserFeedbackHistoryResponse"
        "400":
          description: Invalid user ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to fetch user feedback history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
    Page:
      name: page
      in: query
      description: Page number for pagination (optional - only used when both page and limit are provided)
      schema:
        type: integer
        minimum: 1
    Limit:
      name: limit
      in: query
      description: Number of items per page (optional - only used when both page and limit are provided, max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
  schemas:
...
    ObjectId:
      type: string
      pattern: "^[0-9a-fA-F]{24}$"
      description: MongoDB ObjectId
    Timestamp:
      type: string
      format: date-time
      description: ISO 8601 timestamp

    # Error responses
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Detailed error message

    # Pagination (optional - only when page and limit are both provided)
    PaginationOptional:
      type: object
      properties:
        current:
          type: integer
          description: Current page number
        total:
          type: integer
          description: Total number of pages
        count:
          type: integer
          description: Number of items in current page
        totalRecords:
          type: integer
          description: Total number of records

    # Base list response without pagination
...
    User:
      type: object
      required:
        - _id
        - email
        - createdAt
      properties:
        _id:
          $ref: "#/components/schemas/ObjectId"
        email:
          type: string
          format: email
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
    CreateUserRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
    UpdateUserRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
    CreateUserResponse:
      type: object
      required:
        - message
        - user
      properties:
        message:
          type: string
        user:
          $ref: "#/components/schemas/User"
    UserListResponse:
      type: object
      required:
        - users
        - totalRecords
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        totalRecords:
          type: integer
          description: Total number of users in the collection
        pagination:
          $ref: "#/components/schemas/PaginationOptional"

    # Chat schemas
    SharedUser:
      type: object
      required:
        - userId
        - name
        - email
        - addedAt
      properties:
        userId:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        addedAt:
          $ref: "#/components/schemas/Timestamp"
    Chat:
      type: object
      required:
        - _id
        - id
        - title
        - userId
        - createdAt
        - lastModifiedAt
        - visibility
        - sharedWith
      properties:
        _id:
          $ref: "#/components/schemas/ObjectId"
        id:
          type: string
          description: Custom chat ID
        title:
          type: string
        userId:
          $ref: "#/components/schemas/ObjectId"
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        lastModifiedAt:
          $ref: "#/components/schemas/Timestamp"
        visibility:
          type: string
          enum: [private, shared]
        sharedWith:
          type: array
          items:
            $ref: "#/components/schemas/SharedUser"
    CreateChatRequest:
      type: object
      required:
        - id
        - userId
        - title
      properties:
        id:
          type: string
          description: Custom chat ID
        userId:
          $ref: "#/components/schemas/ObjectId"
        title:
          type: string
...
    CreateChatResponse:
      type: object
      required:
        - message
        - chat
      properties:
        message:
          type: string
        chat:
          $ref: "#/components/schemas/Chat"
...
    ChatListResponse:
      type: object
      required:
        - chats
        - totalRecords
      properties:
        chats:
          type: array
          items:
            $ref: "#/components/schemas/Chat"
        totalRecords:
          type: integer
          description: Total number of chats in the collection
        pagination:
          $ref: "#/components/schemas/PaginationOptional"
...
    Message:
      type: object
      required:
        - _id
        - chatId
        - content
        - createdAt
        - updatedAt
      properties:
        _id:
          $ref: "#/components/schemas/ObjectId"
        chatId:
          type: string
        content:
          type: string
        role:
          type: string
          enum: [user, assistant, system]
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
    CreateMessageRequest:
      type: object
      required:
        - chatId
        - content
      properties:
        chatId:
          type: string
        content:
          type: string
        role:
          type: string
          enum: [user, assistant, system]
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          $ref: "#/components/schemas/Timestamp"
...
    CreateMessageResponse:
      type: object
      required:
        - message
        - data
      properties:
        message:
          type: string
        data:
          $ref: "#/components/schemas/Message"
...
    MessageListResponse:
      type: object
      required:
        - messages
        - totalRecords
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        totalRecords:
          type: integer
          description: Total number of messages in the collection
        pagination:
          $ref: "#/components/schemas/PaginationOptional"

    # Feedback schemas
    JiraIssue:
      type: object
      required:
        - issueKey
        - issueId
        - issueUrl
        - summary
        - description
        - status
        - priority
        - labels
        - createdDate
      properties:
        issueKey:
          type: string
        issueId:
          type: string
        issueUrl:
          type: string
        summary:
          type: string
        description:
          type: string
        status:
          type: string
        assignee:
          type: string
        priority:
          type: string
        labels:
          type: array
          items:
            type: string
        attachments:
          type: array
          items:
            type: string
        createdDate:
          $ref: "#/components/schemas/Timestamp"
    Feedback:
      type: object
      required:
        - _id
        - chatId
        - messageId
        - userId
        - comments
        - rating
        - isUpvoted
        - hasJiraIssue
        - createdAt
        - lastModifiedAt
      properties:
        _id:
          $ref: "#/components/schemas/ObjectId"
        chatId:
          type: string
        messageId:
          type: string
        userId:
          $ref: "#/components/schemas/ObjectId"
        comments:
          type: string
        rating:
          type: integer
          minimum: 0
          maximum: 5
        isUpvoted:
          type: boolean
        hasJiraIssue:
          type: boolean
        jiraIssue:
          $ref: "#/components/schemas/JiraIssue"
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        lastModifiedAt:
          $ref: "#/components/schemas/Timestamp"
...
    FeedbackListResponse:
      type: object
      required:
        - feedbacks
        - totalRecords
      properties:
        feedbacks:
          type: array
          items:
            $ref: "#/components/schemas/Feedback"
        totalRecords:
          type: integer
          description: Total number of feedbacks in the collection
        pagination:
          $ref: "#/components/schemas/PaginationOptional"
    UserFeedbackHistoryResponse:
      type: object
      required:
        - feedbacks
        - totalRecords
      properties:
        feedbacks:
          type: array
          items:
            allOf:
              - $ref: "#/components/schemas/Feedback"
              - type: object
                properties:
                  id:
                    type: string
        totalRecords:
          type: integer
          description: Total number of feedbacks for the user
        pagination:
          allOf:
            - $ref: "#/components/schemas/PaginationOptional"
            - type: object
              properties:
                hasNextPage:
                  type: boolean
                hasPreviousPage:
                  type: boolean

    # Generic response schemas
    UpdateResponse:
      type: object
      required:
        - message
        - modifiedCount
      properties:
        message:
          type: string
        modifiedCount:
          type: integer
    DeleteResponse:
      type: object
      required:
        - message
        - deletedCount
      properties:
        message:
          type: string
        deletedCount:
          type: integer
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
security:
  - bearerAuth: []